ext-diet WP7c+d: EmbodyExt uninstall + settings move to embody_admin module DAT (+6.106 save)
15 methods behind delegating stubs; EmbodyExt 9,412 -> 8,817. Verified
inline (credit-constrained; panels replaced by scripted AST fidelity
checks): 15/15 bodies faithful, marker constants byte-identical at module
scope with zero external readers, deletion-path containment guard observed
refusing an outside-root delete live from the module. 37/37 across four
suites. The 6.106 save ships the module DATs inside the .toe, closing the
frame-5-vs-frame-50 cold-open window found during review (settings restore
fires long before RestoreDATs could recreate missing module DATs on a
stale .toe) -- module-DAT WPs now commit toe+code in lockstep.
