+++ dev/embody/externalizations.tsv
@@ -238,7 +238,7 @@ path	type	strategy	rel_file_path	timestamp	dirty	build	touch_build	node_x	node_y
 /embody/Embody/Collection/CollectionExt	text	py	embody/Embody/Collection/CollectionExt.py	2026-06-20 20:36:05 UTC				200	0	0.9000,0.3000,0.4000
 /embody/Embody/Collection	base	tdn	embody/Embody/Collection.tdn	2026-06-20 20:36:05 UTC		1	099.2025.32820	200	-550	0.3000,0.5000,0.9000
 /embody/Embody/CatalogManagerExt	text	py	embody/Embody/CatalogManagerExt.py	2026-06-20 20:36:05 UTC				-300	300	0.9000,0.3000,0.4000
-/embody/Embody	container	tdn	embody/Embody.tdn	2026-07-06 08:54:46 UTC		742	099.2025.32820	0	0	0.3000,0.5000,0.9000
+/embody/Embody	container	tdn	embody/Embody.tdn	2026-07-06 09:32:22 UTC		742	099.2025.32820	0	0	0.3000,0.5000,0.9000
 /embody	container	tdn	embody.tdn	2026-07-06 08:21:54 UTC		534	099.2025.32820	100	50	0.3000,0.5000,0.9000
 /embody/unit_tests/test_tdn_diff_engine	text	py	embody/unit_tests/test_tdn_diff_engine.py	2026-06-22 21:36:58 UTC				200	-3200	0.9000,0.3000,0.4000
 /embody/unit_tests/test_tdn_diff	text	py	embody/unit_tests/test_tdn_diff.py	2026-06-22 21:37:00 UTC				0	-3200	0.9000,0.3000,0.4000
@@ -268,3 +268,4 @@ path	type	strategy	rel_file_path	timestamp	dirty	build	touch_build	node_x	node_y
 /embody/Embody/envoy_setup	text	py	embody/Embody/envoy_setup.py	2026-07-06 04:19:23 UTC				-500	700	0.9000,0.3000,0.4000
 /embody/Embody/embody_launch	text	py	embody/Embody/embody_launch.py	2026-07-06 08:21:51 UTC				-700	600	0.9000,0.3000,0.4000
 /embody/Embody/embody_git	text	py	embody/Embody/embody_git.py	2026-07-06 08:54:17 UTC				-500	0	0.9000,0.3000,0.4000
+/embody/Embody/embody_admin	text	py	embody/Embody/embody_admin.py	2026-07-06 09:32:19 UTC				-700	400	0.9000,0.3000,0.4000
